+++ lib/decanter/parser/integer_parser.rb
@@ -3,16 +3,10 @@
 module Decanter
   module Parser
     class IntegerParser < ValueParser
-      REGEX = /(\d|[.])/
-
       allow Integer
 
       parser do |val, _options|
-        if val.is_a?(Float)
-          val.to_i
-        else
-          val.scan(REGEX).join.try(:to_i)
-        end
+        Integer(val)
       end
     end
   end